# fixtures/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AuthRole": {
        "description": "Authorization role for API requests (least privilege first).",
        "enum": [
          "reader",
          "operator",
          "admin"
        ],
        "type": "string"
      },
      "AuthTokenCreateRequest": {
        "properties": {
          "label": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/AuthRole"
          }
        },
        "required": [
          "label",
          "role"
        ],
        "type": "object"
      },
      "AuthTokenCreateResponse": {
        "example": {
          "created_at": "2026-04-15T12:00:00Z",
          "label": "ci-deploy",
          "note": null,
          "role": "operator",
          "token_id": "880e8400-e29b-41d4-a716-446655440003",
          "token_prefix": "plct_live_",
          "token_secret": "plct_live_xxxxxxxxxxxxxxxx"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/AuthRole"
          },
          "token_id": {
            "format": "uuid",
            "type": "string"
          },
          "token_prefix": {
            "type": [
              "string",
              "null"
            ]
          },
          "token_secret": {
            "description": "Shown exactly once; never stored in plaintext.",
            "type": "string"
          }
        },
        "required": [
          "token_id",
          "label",
          "role",
          "created_at",
          "token_secret"
        ],
        "type": "object"
      },
      "AuthTokenListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuthTokenSummaryView"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "AuthTokenStatus": {
        "description": "Whether a stored token row is still usable for authentication.",
        "enum": [
          "active",
          "disabled"
        ],
        "type": "string"
      },
      "AuthTokenSummaryView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "disabled_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": "string"
          },
          "last_used_at": {
            "description": "Last successful bearer authentication (updated on each verified request).",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "replaced_by_token_id": {
            "description": "If this row was rotated out, the new token id.",
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/AuthRole"
          },
          "status": {
            "$ref": "#/components/schemas/AuthTokenStatus"
          },
          "token_id": {
            "format": "uuid",
            "type": "string"
          },
          "token_prefix": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "token_id",
          "label",
          "role",
          "status",
          "created_at"
        ],
        "type": "object"
      },
...
      "DataFormat": {
        "description": "Detected or declared on-disk / logical data format.",
        "enum": [
          "parquet",
          "csv",
          "json",
          "ndjson",
          "unknown"
        ],
        "type": "string"
      },
...
      "ErrorBody": {
        "example": {
          "error": {
            "code": "run_not_found",
            "details": {},
            "message": "run not found"
          }
        },
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorEnvelope"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "description": "Wire-level error codes (stable strings).",
        "enum": [
          "invalid_request",
          "invalid_json_request",
          "run_not_found",
          "job_not_found",
          "token_not_found",
          "target_not_found",
          "scan_failed",
          "store_error",
          "internal_error",
          "unauthorized",
          "forbidden",
          "location_not_allowed"
        ],
        "type": "string"
      },
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "JobStatus": {
        "description": "Job lifecycle state persisted in SQLite and returned over HTTP.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "canceled"
        ],
        "type": "string"
      },
...
      "LogicalDatasetTarget": {
        "description": "Logical dataset identifier used when files are resolved indirectly.",
        "properties": {
          "dataset_id": {
            "type": "string"
          },
          "resolver_hint": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "dataset_id"
        ],
        "type": "object"
      },
      "ObjectStoreTargetPlaceholder": {
        "description": "Placeholder for future object-store-backed scans (S3-compatible, GCS, etc.).",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "uri": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "uri"
        ],
        "type": "object"
      },
...
      "RedactionPolicy": {
        "description": "Policy applied before persisting blobs/projections (baseline; expand later).",
        "properties": {
          "strip_finding_evidence_payloads": {
            "description": "Remove `Finding.evidence[].payload` (summaries and locations retained).",
            "type": "boolean"
          },
          "strip_probe_inspection_hints": {
            "description": "Remove `AssetRecord.inspection_hints` before persistence.",
            "type": "boolean"
          },
          "strip_probe_metadata": {
            "description": "Remove `AssetRecord.probe` before persistence.",
            "type": "boolean"
          },
          "truncate_failure_messages_to": {
            "description": "Truncate `failure_message` on assets to this many UTF-8 chars (if set).",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "strip_probe_inspection_hints",
          "strip_probe_metadata",
          "strip_finding_evidence_payloads"
        ],
        "type": "object"
      },
...
      "ScanJobSubmissionResponse": {
        "description": "Response for `POST /api/v1/jobs/scans` and async `POST /api/v1/scans` (202 Accepted).",
        "example": {
          "job_id": "550e8400-e29b-41d4-a716-446655440000",
          "job_url": "/api/v1/jobs/550e8400-e29b-41d4-a716-446655440000",
          "normalized_target_key": "/data/sample.parquet",
          "status": "queued",
          "submitted_at": "2026-04-15T12:00:00Z",
          "target_kind": "local_file"
        },
        "properties": {
          "job_id": {
            "format": "uuid",
            "type": "string"
          },
          "job_url": {
            "type": "string"
          },
          "normalized_target_key": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "submitted_at": {
            "format": "date-time",
            "type": "string"
          },
          "target_kind": {
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "submitted_at",
          "target_kind",
          "normalized_target_key",
          "job_url"
        ],
        "type": "object"
      },
...
      "ScanMode": {
        "description": "How a scan should treat incremental state (placeholder for later phases).",
        "enum": [
          "full",
          "incremental"
        ],
        "type": "string"
      },
      "ScanOptions": {
        "description": "Tunables that influence engine behavior without changing the target identity.",
        "properties": {
          "format_hints": {
            "description": "Optional hint to restrict detection to a subset of formats.",
            "items": {
              "$ref": "#/components/schemas/DataFormat"
            },
            "type": "array"
          },
          "max_files": {
            "description": "Maximum files to enumerate in this scan (safety valve).",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "mode": {
            "$ref": "#/components/schemas/ScanMode"
          },
          "requested_plugins": {
            "description": "Explicitly requested analyzer plugins (empty = run none).",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ScanProfile": {
        "description": "Coarse scan depth presets.",
        "enum": [
          "quick",
          "standard",
          "deep",
          "baseline"
        ],
        "type": "string"
      },
...
      "ScanTarget": {
        "description": "Concrete scan target with mutually consistent fields.",
        "oneOf": [
          {
            "properties": {
              "path": {
                "type": "string"
              },
              "type": {
                "enum": [
                  "local_file"
                ],
                "type": "string"
              }
            },
            "required": [
              "path",
              "type"
            ],
            "type": "object"
          },
          {
            "properties": {
              "path": {
                "type": "string"
              },
              "type": {
                "enum": [
                  "local_directory"
                ],
                "type": "string"
              }
            },
            "required": [
              "path",
              "type"
            ],
            "type": "object"
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/LogicalDatasetTarget"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "logical_dataset"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectStoreTargetPlaceholder"
              },
              {
                "properties": {
                  "type": {
                    "enum": [
                      "object_store_placeholder"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
...
      "StartScanRequest": {
        "properties": {
          "options": {
            "$ref": "#/components/schemas/ScanOptions"
          },
          "profile": {
            "$ref": "#/components/schemas/ScanProfile"
          },
          "redaction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RedactionPolicy",
                "description": "When omitted, [`parqonaut_types::RedactionPolicy::transport_safe_persist`] is applied before persistence."
              }
            ]
          },
          "scan_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "target": {
            "$ref": "#/components/schemas/ScanTarget"
          }
        },
        "required": [
          "target",
          "profile"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Opaque API token; set `Authorization: Bearer <token>`. Tokens are stored hashed in SQLite.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Joshua Grant"
    },
    "description": "PARQONAUT HTTP API for scan jobs, persisted runs, diffing, and bearer-token administration. Successful JSON responses and application-level errors use the `{ \"error\": { \"code\", \"message\", \"details\" } }` envelope for failures. JSON request bodies use a shared extractor so malformed JSON, wrong JSON shape, missing `Content-Type: application/json`, and related body-buffer failures map to `invalid_json_request` with appropriate HTTP status codes (`400`, `415`, `422`, `413` where applicable).",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "PARQONAUT API",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/tokens": {
      "get": {
        "operationId": "get_admin_tokens",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenListResponse"
                }
              }
            },
            "description": "All API tokens (metadata only; no secrets)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Admin role required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "post_admin_tokens",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthTokenCreateRequest"
              }
            }
          },
          "description": "Label, role, optional note",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenCreateResponse"
                }
              }
            },
            "description": "Token created; secret returned once"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid JSON request body (`invalid_json_request`; e.g. syntax, body buffer)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Admin role required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request body exceeded configured limit while buffering (`invalid_json_request`)"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or non-JSON `Content-Type` (`invalid_json_request`)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "JSON value did not match the expected request shape (`invalid_json_request`)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/scans": {
      "post": {
        "operationId": "post_scans_async",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartScanRequest"
              }
            }
          },
          "description": "Scan target, profile, and options",
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobSubmissionResponse"
                }
              }
            },
            "description": "Async scan job accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid JSON request body (`invalid_json_request`; e.g. syntax, body buffer)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Insufficient role (operator required)"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request body exceeded configured limit while buffering (`invalid_json_request`)"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or non-JSON `Content-Type` (`invalid_json_request`)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "JSON value did not match the expected request shape (`invalid_json_request`)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}